Review feedback: Rename StreamsCache -> StreamsRecordCache
